fix(codec-opus): SILK decode correctness, tell() semantics, CELT hardening
Reopening #313-#316 found the decoder producing plausible-RMS noise
rather than a real decode. Chased with the RFC 6716 test vectors
(fetched from opus-codec.org, sha1-verified against the RFC's own
Appendix A.4 list) and a locally built, lightly instrumented libopus
(BSD, Tier A per planning/research/07-legal-patents-licensing.md) as a
bit-for-bit oracle -- range-decoder final-range fields in the .bit
files and printf tracing of libopus's own intermediate values, never
its source pasted into this tree.

SILK: nine real bugs, found by diffing intermediate values against the
traced reference for the same frames.

1. `silk/nlsf.rs` `decode_nlsf_indices`: the stage-1 codebook's
   voiced/unvoiced row stride used `cb.order` (10 or 16) instead of
   `cb.n_vectors` (32), so every voiced frame's NLSF stage-1 index read
   from the wrong offset into the table.
2. `silk/decode.rs` `shell_decode`: the 16-pulse shell tree was decoded
   breadth-first (all of level 2, then all of level 1, then all of
   level 0) instead of the reference's depth-first traversal order.
   Every `decode_split` call reads from the shared range coder, so the
   wrong order reads each pulse-count split against the wrong context.
3. `silk/decode.rs` sign decoding: the table-row selector had
   `quantOffsetType` and `signalType` swapped
   (`2*quantOffsetType+signalType` instead of the reference's
   `quantOffsetType+2*signalType`), and was missing the reference's
   `min(p & 0x1F, 6)` clamp on the per-block row index.
4. `silk/nlsf.rs` `find_poly`: the P/Q polynomial builder indexed its
   already-pre-strided (every-other-sample) input array with `2*k`
   again, walking off the end for the back half of every band and
   silently substituting zero there.
5. `silk/nlsf.rs` `nlsf_to_lpc`: the P/Q-to-`a` combination was missing
   the factor of two `silk_NLSF2A`'s combination step produces (`a32`
   comes out one bit wider than `P`/`Q`'s own scale) -- LPC coefficients
   were exactly 2x too large.
6. `silk/decode.rs` `synthesize`: the AR/LTP recursion divided its
   already real-valued `a` coefficients (from `nlsf_to_lpc`, not a raw
   Q12 integer) by 4096 again, making the short-term predictor
   contribute almost nothing.
7. `silk/decode.rs`: `first_frame_after_reset` was set `true` at
   construction and never cleared, so NLSF interpolation was force-
   disabled for the life of the decoder instead of only the first
   frame.
8. `silk/decode.rs` `synthesize`: LTP re-whitening ran on every
   subframe instead of only subframe 0, or subframe 2 when NLSF
   interpolation is active (`decode_core.c`'s actual re-whitening
   cadence) -- rewritten to a persistent `ltp_old`/`ltp_new` delay line
   matching `sLTP_Q15`, rescaled by the gain ratio on other subframes
   rather than re-derived from `out_buf`.
9. `silk/mod.rs` `SilkDecoder::decode`: `nb_subfr` was fixed at
   construction/reconfigure time; `ensure_silk` only reconfigures on an
   internal-rate change, so a stream switching frame duration (e.g. 20ms
   to 10ms) at a fixed rate decoded every following frame with a stale
   subframe count. Now recomputed from `frame_ms` every call.

Measured against the RFC vectors' 12kHz-internal SILK-only content
(testvector03, mono segment): range-decoder final-range now matches
the reference for all 510 packets (was failing from packet 0), and
end-to-end `OpusDecoder` output correlates ~0.58 against the reference
across the whole segment (was ~0, indistinguishable from noise).

`range.rs`: `RangeDecoder::tell()` computed `tell_frac() >> BITRES`
instead of the reference's own distinct `ec_tell()` formula
(`nbits_total - EC_ILOG(rng)`); the two are close but not identical,
and CELT's frame syntax uses `tell() + N <= budget` to decide whether
an optional field is present, so a decoder whose estimate disagrees at
exactly the wrong boundary makes a different presence decision than
the bitstream was encoded with. Confirmed against the reference trace:
this alone corrected a real coarse-energy divergence in one CELT
frame. Return type changed from `u32` to `i32` to match the
reference's signed `int` (can be non-positive immediately after
construction); all call sites already cast to `i32`.

`celt/energy.rs`: `unquant_fine_energy` was missing the reference's
budget guard (skip a band whose raw bits would run past the frame's
byte budget) before reading them; and `unquant_coarse_energy`'s
per-band energy update had no bound, where the reference clamps it to
+-28dB but only inside `#ifdef FIXED_POINT` (a no-op on the reference's
own float build). `laplace_decode`'s escape mechanism can still return
an arbitrarily large value given enough adversarial input bits, which
`log2_amp`'s `exp2` turns into `inf`/`NaN`; a fuzz run
(`opus_decode`, 60s) found this via the harness's own "every sample is
finite" invariant. Clamping unconditionally closes it without
affecting real content. Fuzz artifact diagnosed and fixed, not just
deleted; `find fuzz/artifacts -type f` is empty for this target.

Verified with the CELT range-state check against testvector07
(CELT-only): coarse energy, `tf_decode`, spread decision, dynalloc,
trim and the full bit allocation are now bit-exact against the traced
reference for a representative transient frame that previously
desynced. The remaining CELT gap is narrower than before but not
closed: the PVQ/short-block band-splitting recursion still diverges
under transient short-block conditions for some frames (roughly half
of testvector07 still fails the range-state check) -- `haar1`, the
Hadamard interleave/deinterleave primitives, the `tf_select` table and
`compute_qn` are all verified bit-exact, so the remaining defect is
somewhere in the recursive theta-split/allocation logic in
`bands.rs`'s `quant_band`, not yet localized further.

The decoder remains unregistered (see the earlier unregister commit)
pending that fix and a hybrid re-measurement, per the reopened issues'
stop condition. Not closing #313/#314/#315/#316: SILK is now
demonstrably correct on the tested content, but CELT and hybrid are
not there yet.

Vaco-Spec-Ref: rfc-6716 opus_testvectors (opus-codec.org, sha1-verified against RFC 6716 Appendix A.4) plus a locally built libopus (BSD, Tier A) used as a black-box/instrumented oracle -- no libopus source copied into this tree
Vaco-Provenance: rfc
Vaco-Clean-Room: yes
Vaco-AI-Assisted: yes

+++ crates/codec/vaco-codec-opus/src/celt/mod.rs
@@ -140,7 +140,7 @@ impl CeltDecoder {
         let n = m * 120;
 
         let total_bits_bytes = len_bytes as i32 * 8;
-        let mut tell = dec.tell() as i32;
+        let mut tell = dec.tell();
 
         let silence = if tell >= total_bits_bytes {
             true
@@ -159,16 +159,16 @@ impl CeltDecoder {
                 let octave = dec.dec_uint(6).unwrap_or(0);
                 let _period = (16u32 << octave) + dec.dec_bits(4 + octave) - 1;
                 let _qg = dec.dec_bits(3);
-                if dec.tell() as i32 + 2 <= total_bits_bytes {
+                if dec.tell() + 2 <= total_bits_bytes {
                     let _tapset = dec.icdf(&tables::TAPSET_ICDF, 2).unwrap_or(0);
                 }
             }
-            tell = dec.tell() as i32;
+            tell = dec.tell();
         }
 
         let is_transient = if lm > 0 && tell + 3 <= total_bits_bytes {
             let v = dec.bit_logp(3);
-            tell = dec.tell() as i32;
+            tell = dec.tell();
             v
         } else {
             false
@@ -181,7 +181,7 @@ impl CeltDecoder {
         let mut tf_res = vec![0i32; NB_EBANDS];
         tf_decode(dec, start_band, end_band, is_transient, &mut tf_res, lm, total_bits_bytes);
 
-        tell = dec.tell() as i32;
+        tell = dec.tell();
         let spread = if tell + 4 <= total_bits_bytes { dec.icdf(&tables::SPREAD_ICDF, 5).unwrap_or(2) } else { 2 };
 
         let cap = rate::init_caps(&EBANDS, lm, channels as i32);
@@ -248,7 +248,7 @@ impl CeltDecoder {
 
         let anti_collapse_on = if anti_collapse_rsv > 0 { dec.dec_bits(1) != 0 } else { false };
 
-        let bits_left = len_bytes as i32 * 8 - dec.tell() as i32;
+        let bits_left = len_bytes as i32 * 8 - dec.tell();
         energy::unquant_energy_finalise(
             dec,
             &mut self.old_band_e,
@@ -390,7 +390,7 @@ impl CeltDecoder {
 
 /// `celt.c`'s `tf_decode`.
 fn tf_decode(dec: &mut RangeDecoder<'_>, start: usize, end: usize, is_transient: bool, tf_res: &mut [i32], lm: i32, total_bits: i32) {
-    let mut tell = dec.tell() as i32;
+    let mut tell = dec.tell();
     let mut logp = if is_transient { 2 } else { 4 };
     let tf_select_rsv = i32::from(lm > 0 && tell + logp < total_bits);
     let budget = total_bits - tf_select_rsv;
@@ -400,7 +400,7 @@ fn tf_decode(dec: &mut RangeDecoder<'_>, start: usize, end: usize, is_transient:
         if tell + logp <= budget {
             let bit = dec.bit_logp(logp as u32);
             curr ^= bit;
-            tell = dec.tell() as i32;
+            tell = dec.tell();
             tf_changed |= curr;
         }
         *slot = i32::from(curr);
